+++ src/main/java/whz/pti/models/DeviceType.java
@@ -1,25 +1,23 @@
 package whz.pti.models;
 
-import java.math.BigDecimal;
-
 public class DeviceType {
-    private BigDecimal id;
+    private Long id;
     private String name;
     private String description;
 
     public DeviceType() {}
 
-    public DeviceType(BigDecimal id, String name, String description) {
+    public DeviceType(Long id, String name, String description) {
         this.id = id;
         this.name = name;
         this.description = description;
     }
 
-    public BigDecimal getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(BigDecimal id) {
+    public void setId(Long id) {
         this.id = id;
     }
